Update, add attributor, and fix batch lookups
I add the attributor. But before I could do that, intellij notified me that my dependencies needed to be updated, so I did that, which broke a bunch of things. Then when I tried to test the attributor, I found that a missing mod was completely breaking my batch lookup system, so I fixed that too.